+++ core/store/src/trie/raw_node.rs
@@ -41,13 +41,6 @@ impl RawTrieNode {
             None => Self::BranchNoValue(children),
         }
     }
-
-    pub fn has_value(&self) -> bool {
-        match self {
-            Self::BranchWithValue(_, _) | Self::Leaf(_, _) => true,
-            _ => false,
-        }
-    }
 }
 
 /// Children of a branch node.
