+++ tools/state-viewer/src/contract_accounts.rs
@@ -8,7 +8,8 @@ use near_primitives::transaction::{Action, ExecutionOutcomeWithProof};
 use near_primitives::trie_key::trie_key_parsers::parse_account_id_from_contract_code_key;
 use near_primitives::trie_key::TrieKey;
 use near_primitives::types::AccountId;
-use near_store::{DBCol, NibbleSlice, StorageError, Store, Trie, TrieTraversalItem};
+use near_store::trie::ops::iter::TrieTraversalItem;
+use near_store::{DBCol, NibbleSlice, StorageError, Store, Trie};
 use std::collections::{BTreeMap, BTreeSet, VecDeque};
 
 type Result<T> = std::result::Result<T, ContractAccountError>;
